added: jlBRDF for reflectante functions jlBRDFLambertian for reflectante lamber jlMaterial for any kind of material jlMMatte for a matte material jlLights for any kind of light jlLAmbient is a class for ambient lights jlPointLight class for point lighrs jlTRayCast, a new tracer, it can get the color calculating lights with object's materials
Added imgui and openGL for interface

add 2 new functions of traceRay in jlTracer to trace the rays with depth, one of them is used for jlTRayCast
The objects have material now
jlShadeRec now save the data of objects and materials for calculate the light influence

+++ jdLightning/jdLightning/src/jlGOSphere.cpp
@@ -24,15 +24,15 @@ jlSphere::hit(const jlRay& ray, double& tmin, jlShadeRec& sr) {
     t = (-b - e) / denom; // smaller root
     if (t > kEpsilon) {
       tmin = t;
-      sr.m_normal = (temp + t * ray.m_direction) / m_radius;
-      sr.m_localHitPoint = ray.m_origin + t * ray.m_direction;
+      sr.m_normal = (temp + (float)t * ray.m_direction) / m_radius;
+      sr.m_localHitPoint = ray.m_origin + (float)t * ray.m_direction;
       return true;
     }
     t = (-b + e) / denom; // larger root
     if (t > kEpsilon) {
       tmin = t;
-      sr.m_normal = (temp + t * ray.m_direction) / m_radius;
-      sr.m_localHitPoint = ray.m_origin + t * ray.m_direction;
+      sr.m_normal = (temp + (float)t * ray.m_direction) / m_radius;
+      sr.m_localHitPoint = ray.m_origin + (float)t * ray.m_direction;
       return true;
     }
   }
